+++ multi-target-decision-tree/src/utils/calculations/calculations.rs
@@ -5,18 +5,18 @@ pub mod variance {
 
     #[derive(Debug)]
     pub struct MultiTargetLabelMetrics {
-        pub sum_of_labels_vector: Vec<f32>,
-        pub sum_of_squared_labels_vector: Vec<f32>,
-        pub mean_of_labels_vector: Vec<f32>,
+        pub sum_of_labels_vector: Vec<f64>,
+        pub sum_of_squared_labels_vector: Vec<f64>,
+        pub mean_of_labels_vector: Vec<f64>,
     }
 
     pub fn calculate_loss_vector(
-        left_variance_vector: Vec<f32>,
-        right_variance_vector: Vec<f32>,
-        left_size: f32,
-        right_size: f32,
+        left_variance_vector: Vec<f64>,
+        right_variance_vector: Vec<f64>,
+        left_size: f64,
+        right_size: f64,
         number_of_targets: usize,
-    ) -> Vec<f32> {
+    ) -> Vec<f64> {
         calculations_variance::calculate_loss_vector(
             left_variance_vector,
             right_variance_vector,
@@ -28,9 +28,9 @@ pub mod variance {
 
     pub fn calculate_variance_vector(
         multi_target_label_metrics: &MultiTargetLabelMetrics,
-        number_of_labels: f32,
+        number_of_labels: f64,
         number_of_targets: usize,
-    ) -> Vec<f32> {
+    ) -> Vec<f64> {
         calculations_variance::calculate_variance_vector(
             multi_target_label_metrics,
             number_of_labels,
@@ -39,7 +39,7 @@ pub mod variance {
     }
 
     pub fn get_multi_target_label_metrics(
-        labels: &Vec<Vec<f32>>,
+        labels: &Vec<Vec<f64>>,
         number_of_targets: usize,
     ) -> MultiTargetLabelMetrics {
         calculations_variance::get_multi_target_label_metrics(labels, number_of_targets)
